### 2025-04-10 #697 @n-kall
Presubmission inquiry: priorsense - Prior Diagnostics and Sensitivity Analysis
Submitting Author Name: Noa Kallioinen
Submitting Author Github Handle: @n-kall
Other Package Authors Github handles: (comma separated, delete if none) @topipa, @paul-buerkner, @avehtari
Repository:  https://github.com/n-kall/priorsense
Submission type: Pre-submission

---

```
Package: priorsense
Title: Prior Diagnostics and Sensitivity Analysis
Version: 1.1.1.9000
Authors@R: c(person("Noa", "Kallioinen", email = "noa.kallioinen@aalto.fi", role = c("aut", "cre", "cph")),
	     person("Topi", "Paananen", role = c("aut")),
	     person("Paul-Christian", "Bürkner", role = c("aut")),
	     person("Aki", "Vehtari", role = c("aut")),
	     person("Frank", "Weber", role = c("ctb"))
	     )
Description: Provides functions for prior and likelihood sensitivity analysis in Bayesian models. Currently it implements methods to determine the sensitivity of the posterior to power-scaling perturbations of the prior and likelihood.
Roxygen: list(markdown = TRUE, roclets = c ("namespace", "rd", "srr::srr_stats_roclet"))
    checkmate (>= 2.3.1),
    ggdist (>= 3.3.2),
    ggh4x (>= 0.2.5),
    ggplot2 (>= 3.5.1),
    grDevices (>= 3.6.2),
    matrixStats (>= 1.3.0),
    posterior (>= 1.6.0),
    rlang (>= 1.1.4),
    tibble (>= 3.2.1),
    utils
    R2jags (>= 0.8),
    bayesplot (>= 1.11.1),
    brms (>= 2.22.0),
    cmdstanr (>= 0.8.1),
    iwmm (>= 0.0.1),
    quarto,
    philentropy (>= 0.8.0),
    rstan (>= 2.32.6),
    transport (>= 0.15)
VignetteBuilder: quarto
Additional_repositories:
    https://topipa.r-universe.dev,
    https://stan-dev.r-universe.dev
Config/Needs/website: quarto
```

- priorsense provides diagnostic tools for checking the prior and likelihood sensitivity of Bayesian models. While it does not generate posterior draws itself, it is part of the greater Bayesian workflow and seems to fit best in "Bayesian and Monte Carlo Routines"

- Answer: Yes, srr standards are implemented.

-   Answer: Users of Bayesian models who would like to check the prior (and likelihood) sensitivity of their model.

-   Answer: No other R packages implement the method.

-  Answer: not applicable

-  Answer: Many of the standards for "Bayesian and Monte Carlo Routines" focus on software that performs sampling from the posterior, which priorsense does not do. So these standards were deemed non-applicable.
