+++ pkgdown.yml
@@ -7,7 +7,7 @@ articles:
   quantity_of_interest: quantity_of_interest.html
   selecting_priors_and_quantities: selecting_priors_and_quantities.html
   sensitivity_diagnostic: sensitivity_diagnostic.html
-last_built: 2025-04-10T12:59Z
+last_built: 2025-04-10T13:04Z
 urls:
   reference: https://n-kall.github.io/priorsense/reference
   article: https://n-kall.github.io/priorsense/articles
