### 2025-04-11 81e1c8b
add tests with a simple conjugate model
