### 2025-04-11 #697 @mpadge
Thanks for inquiring @n-kall, it looks like an interesting package that should be able to be reviewed. The checks nevertheless reveal a couple of issues we'll first need to address:

### R CMD check failures

You'll see in the report above that your package fails on our check systems because we don't have `cmdstanr` properly configured. I think that package is sufficiently important for us to have it pre-installed and configured on our systems. As explained in [our *Dev Guide*](https://devguide.ropensci.org/softwarereview_author.html#:~:text=unusual%20system%20dependencies), could you please submit a pull request to [our base Dockerfile](https://github.com/ropensci-review-tools/pkgcheck/blob/main/Dockerfile) to update that.

It might help you first to use our [`pkgcheck-action`](https://github.com/ropensci-review-tools/pkgcheck-action) in your own repo to experiment and confirm. You can add the additional steps under `jobs:` -> `steps:` in that workflow file to confirm that it all works there. I see you've already got the relevant lines in your own rcmdcheck workflow, so should be pretty straightforward. When preparing your PR to our Dockerfile, you can add the environment variable at the end of [these lines](https://github.com/ropensci-review-tools/pkgcheck/blob/1b80b1c3642f0c50afae8a97ecd1d8cbde353629/Dockerfile#L256-L258).

### srr failures

Please note that the `srr` package has been recently updated. With the latest version (0.1.4.002 or later), you should see this:

``` r
srr::srr_stats_pre_submit()
#> ! Package can not be submitted because the following standards [v0.2.0] are missing from your code:
#> 1. G5.4a
#> 2. G5.5
#> 3. G5.6
#> 4. G5.6a
#> 5. G5.6b
#> 6. G5.7
#> 7. G5.9
#> 8. G5.9b
#> ! Package must comply with at least 50% of category-specific standards, but currently complies with only 24%
```

The last statement is particularly important. You currently comply with only 13 out of our 56 standards for Bayesian software. Could you please return to those standards, and endeavour to comply with at least half of them? These pre-submission issues are intended to allow all kinds of discussions on getting packages up to scratch, so please feel free to use this issue to ask any questions you might have while you increase compliance with our standards.

### Additional issue

A minor additional issues appears within the "*goodpractice and other checks*" section above, where you'll see a long list of `lintr` objections to long lines. It would be good if you could apply `lintr` directly to your package to reduce the output noise generated there.

---

Thanks again for your inquiry. Looking forward to working with you to help transition your package towards a full submission.
