### 2025-04-11 #697 @n-kall
@mpadge Thanks for your helpful response!
- I had forgotten to push one test file which addresses the missing general standards. Those should now be fixed.
- I can make a PR to include `cmdstanr` in the base dockerfile
- Regarding missing standards for Bayesian software, I noted in my initial post that many of these standards assume that the software package does posterior estimation via sampling. The standards refer to specifying prior distributions, data, sampling algorithms, convergence checks. As `priorsense` does not do any sampling, these did not seem applicable (and is documented in `@ssrstatsNA` comments).
- Perhaps "Bayesian and Monte Carlo software" is not the correct category, as I now see that the category "Exploratory data analysis and summary statistics" also includes exploring models. Although `priorsense` is explicitly for checking Bayesian models and has similar functionality to the first two examples in the "Bayesian and Monte Carlo software" (ArviZ and bayestestR), maybe I should change the category?
