+++ tests/testthat/test_powerscale.R
@@ -1,5 +1,8 @@
+#' @srrstats {EA6.0} return values are tested in tests
+
 univariate_normal_draws <- example_powerscale_model()$draws
 
+#' @srrstats {EA4.0} output type tested
 test_that("priorsense_data is created", {
   expect_s3_class(
     create_priorsense_data(
@@ -11,6 +14,7 @@ test_that("priorsense_data is created", {
 )
 
 #' @srrstats {G5.3} Missing values tested
+#' @srrstats {EA4.0} output type tested
 test_that("powerscale returns powerscaled_draws with no missing values", {
   psp <- powerscale(
       x = univariate_normal_draws,
@@ -38,6 +42,7 @@ test_that("powerscale returns powerscaled_draws with no missing values", {
 }
 )
 
+#' @srrstats {EA4.0} output type tested
 test_that("powerscale_seqence returns powerscaled_sequence", {
   expect_s3_class(
     suppressWarnings(powerscale_sequence(
@@ -48,13 +53,26 @@ test_that("powerscale_seqence returns powerscaled_sequence", {
 }
 )
 
-test_that("powerscale_sensitivity returns powerscaled_sensitivity_summary", {
-  expect_s3_class(
-    powerscale_sensitivity(
+#' @srrstats {EA6.0a} classes tested
+#' @srrstats {EA6.0b} dimensions tested
+#' @srrstats {EA6.0c} column names tested
+#' @srrstats {EA6.0d} classes of columns tested
+test_that("powerscale_sensitivity returns powerscaled_sensitivity_summary with expected columns", {
+  ps <- powerscale_sensitivity(
       x = univariate_normal_draws
-    ),
+    )
+  expect_s3_class(ps
+,
     "powerscaled_sensitivity_summary"
   )
+
+  expect_equal(dim(ps), c(2, 4))
+  
+  expect_identical(colnames(ps), c("variable", "prior", "likelihood", "diagnosis"))
+  expect_vector(ps[["variable"]], ptype = character())
+  expect_vector(ps[["prior"]], ptype = numeric())
+  expect_vector(ps[["likelihood"]], ptype = numeric())
+  expect_vector(ps[["diagnosis"]], ptype = character())
 }
 )
 
@@ -86,7 +104,9 @@ test_that("powerscale_sequence uses input alphas correctly", {
   )
 
   expect_equal(
-    get_powerscaling_details(pss$prior_scaled$draws_sequence[[length(pss$alphas)]])$alpha,
+    get_powerscaling_details(
+      pss$prior_scaled$draws_sequence[[length(pss$alphas)]]
+    )$alpha,
     2.5
   )
 
@@ -106,12 +126,17 @@ test_that("powerscale_sequence adapts alphas and keeps pareto-k low", {
   ))
 
   expect_lt(
-    get_powerscaling_details(pss$likelihood_scaled$draws_sequence[[1]])$diagnostics$khat,
+    get_powerscaling_details(
+      pss$likelihood_scaled$draws_sequence[[1]]
+    )$diagnostics$khat,
     k_threshold
   )
 
   expect_lt(
-    get_powerscaling_details(pss$likelihood_scaled$draws_sequence[[length(pss$likelihood_scaled$draws_sequence)]])$diagnostics$khat,
+    get_powerscaling_details(
+      pss$likelihood_scaled$draws_sequence[[length(
+        pss$likelihood_scaled$draws_sequence)]]
+    )$diagnostics$khat,
     k_threshold
   )
 
@@ -121,7 +146,9 @@ test_that("powerscale_sequence adapts alphas and keeps pareto-k low", {
   )
 
   expect_lt(
-    get_powerscaling_details(pss$prior_scaled$draws_sequence[[length(pss$prior_scaled$draws_sequence)]])$diagnostics$khat,
[…]
