### 2025-04-14 43f8252
remove bayesian statistics category
