+++ R/srr-stats-standards.R
@@ -47,42 +47,6 @@ NULL
 #' @srrstatsNA {G5.11a} No downloads are required for tests.
 #' @srrstatsNA {G2.12} tabular objects are handled via the `posterior`
 #'   package
-#' @srrstatsNA {BS1.0} The term "hyperparameter" is not used.
-#' @srrstatsNA {BS1.1, BS1.2, BS1.2a, BS1.2b, BS1.2c} Model fitting is
-#'   not part of the software package, the user should provide the
-#'   posterior draws that are generated via some other means.
-#' @srrstatsNA {BS1.3, BS1.3a, BS1.3b} No sampling algorithm is
-#'   implemented, users should provide already generated posterior
-#'   draws.
-#' @srrstatsNA {BS1.4, BS1.5} No convergence checkers are implemented,
-#'   as they are not needed.
-#' @srrstatsNA {BS2.1, BS2.1a} No data is used, instead only the
-#'   posterior draws
-#' @srrstatsNA {BS2.2, BS2.3, BS2.4, BS2.5} No distributional
-#'   parameters are handled in the software package, the model should
-#'   be fit with some other software beforehand
-#' @srrstatsNA {BS2.7, BS2.8, BS2.9, BS2.10, BS2.11} No sampling
-#'   algorithm is implemented, users should provide already generated
-#'   posterior draws.
-#' @srrstatsNA {BS3.1, BS3.2} Perfect collinearity is not an issue for
-#'   prior diagnostics
-#' @srrstatsNA {BS4.6, BS4.7} No convergence checker implemented or
-#'   needed
-#' @srrstatsNA {BS4.0, BS4.1, BS4.2, BS4.3, BS4.4, BS4.5} The software
-#'   does to generate draws from the posterior distribution, it
-#'   analyses already generated ones
-#' @srrstatsNA {BS5.0} The software does not generate draws requiring
-#'   a seed or initial state
-#' @srrstatsNA {BS5.1} Model fitting is not performed, so input data
-#'   structure is not relevant
-#' @srrstatsNA {BS5.4} No convergence checkers implemented or needed
-#' @srrstatsNA {BS6.2} The software package does not create sequences
-#'   of posterior draws
-#' @srrstatsNA {BS6.5} The software does not create sequences of
-#'   posterior draws
-#' @srrstatsNA {BS7.0, BS7.1, BS7.2, BS7.3, BS7.4, BS7.4a} The
-#'   software does not perform model fitting, so parameter recovery
-#'   and prediction are not relevant
 #'
 #' @srrstatsNA {EA5.6} No bundled libraries
 #' @srrstatsNA {EA5.6} No bundled libraries
