### 2025-04-15 eb3d888
description and comment cleanup
