+++ DESCRIPTION
@@ -26,13 +26,13 @@ Imports:
     tibble (>= 3.2.1),
     utils
 Suggests:
-    R2jags (>= 0.8),
     bayesplot (>= 1.11.1),
     brms (>= 2.22.0),
     cmdstanr (>= 0.8.1),
     iwmm (>= 0.0.1),
-    quarto,
     philentropy (>= 0.8.0),
+    quarto (>= 1.4.4),
+    R2jags (>= 0.8),
     rstan (>= 2.32.6),
     testthat (>= 3.0.0),
     transport (>= 0.15),
