### 2025-04-15 f7eff8d
add pkgcheck action
