+++ .github/workflows/pkgcheck.yaml
@@ -1,31 +1,0 @@
-name: pkgcheck
-
-# This will cancel running jobs once a new run is triggered
-concurrency:
-  group: ${{ github.workflow }}-${{ github.head_ref }}
-  cancel-in-progress: true
-
-on:
-  # Manually trigger the Action under Actions/pkgcheck
-  workflow_dispatch:
-  # Run on every push to main
-  push:
-    branches:
-      - main
-      - development
-
-jobs:
-  pkgcheck:
-    runs-on: ubuntu-latest
-    permissions:
-      issues: write
-    steps:
-      - name: set cmdstan path
-        run: |
-          echo "CMDSTAN_PATH=${HOME}/.cmdstan" >> $GITHUB_ENV
-      - name: Install cmdstan
-        run: |
-          cmdstanr::check_cmdstan_toolchain(fix = TRUE)
-          cmdstanr::install_cmdstan(cores = 2)
-        shell: Rscript {0}
-      - uses: ropensci-review-tools/pkgcheck-action@main
