### 2025-04-23 #697 @n-kall
- I have changed the category to "Exploratory data analysis and summary statistics" and now fulfill at least half of the standards from that category.
- I have also made the PR to add cmdstan (and rstan) to the docker image

Can the package be rechecked?
