### 2025-04-25 #697 @mpadge
Thanks @n-kall. Requests to `@ropensci-review-bot` have to be individual comments. Like the following one which I'll issue now, and leave the full checks up to you.
