### 2025-05-04 #697 @maurolepore
This marks the end of my EiC rotation and leaver a comment for the next EiC.

* This is a stats pre-submission
* @mpadge is guiding the process and I see good communication

@n-kall thanks for sharing your work with rOpenSci!
