### 2025-05-05 #697 @n-kall
Thanks for the guidance so far with this pre-submission @mpadge. Now that the package is in "top shape", is there anything I need to do now to transition it to a full submission for review?
